+++ lib/src/commands/upstream_branch.dart
@@ -6,6 +6,7 @@
 
 import 'dart:io';
 
+import 'package:gg_args/gg_args.dart';
 import 'package:gg_git/src/base/gg_git_base.dart';
 import 'package:gg_log/gg_log.dart';
 
@@ -67,3 +68,7 @@ class UpstreamBranch extends GgGitBase<String> {
     }
   }
 }
+
+/// Mocktail mock
+class MockUpstreamBranch extends MockDirCommand<String>
+    implements UpstreamBranch {}