+++ lib/src/test_helpers/test_helpers.dart
@@ -190,6 +190,14 @@ Future<String> branchName(Directory testDir) async {
   return result.stdout.toString().trim();
 }
 
+// .............................................................................
+/// Returns the name of the upstream branch
+Future<String> upstreamBranchName(Directory testDir) async {
+  final result =
+      UpstreamBranch(ggLog: print).get(directory: testDir, ggLog: print);
+  return result;
+}
+
 // ######################
 // Git Ignore
 // ######################