+++ test/commands/commit_test.dart
@@ -177,7 +177,7 @@ void main() {
           );
         });
 
-        test('if ammand and ammendWhenNotPushed is true at the same time',
+        test('if ammend and ammendWhenNotPushed is true at the same time',
             () async {
           // Let's modify a file
           await addFileWithoutCommitting(d, fileName: 'file1.txt');
@@ -377,6 +377,57 @@ void main() {
                 directory: d,
               );
               expect(commitMessage2, 'Commit message 2');
+            },
+          );
+        });
+
+        group('when ammendWhenNotPushed is true and branch has no remote', () {
+          test(
+            'and it should not overwrite the last commit message',
+            () async {
+              // Create a new local branch which does not has a remote branch
+              await createBranch(d, 'feature1');
+
+              // Let's modify a file
+              await addAndCommitSampleFile(
+                d,
+                fileName: 'file1.txt',
+                message: 'Commit message 0',
+              );
+
+              // Count the number of commits
+              final count0 = await commitCount.get(
+                ggLog: messages.add,
+                directory: d,
+              );
+
+              // Modify the file again
+              File('${d.path}/file1.txt').writeAsStringSync('Change 2!');
+
+              // Commit the file again with ammendWhenNotPushed = true
+              await commit.commit(
+                directory: d,
+                message: 'Commit message 1',
+                doStage: true,
+                ammendWhenNotPushed: true,
+                ggLog: messages.add,
+              );
+
+              // Commit count should not have increased
+              // because we did not push yet.
+              final count1 = await commitCount.get(
+                ggLog: messages.add,
+                directory: d,
+              );
+              expect(count1, count0);
+
+              // Commit message should be the previous one.
+              // because the commit was ammended
+              final commitMessage1 = await headMessage.get(
+                ggLog: messages.add,
+                directory: d,
+              );
+              expect(commitMessage1, 'Commit message 0');
             },
           );
         });